Add move testing utilities and update imports
